pyproject.toml
[project]
name = "starpipe-simple"
version = "0.1.0"
description = "Simple pipecat voice AI tutor"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[openai,elevenlabs,daily]>=0.0.50",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "onnxruntime>=1.19.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
